Various loop optimizations, mostly matters on gameplay classes.
Ugly but does actually make a difference in heavy mods.
Mostly important on critical gameplay executions but did it on most classes just because I can.

Goal here is not nice code but to squeeze as much as we can out of Psych Engines codebase as much as we can without breaking existing functionality.